% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBET.R
\name{item_characteristics_sampler_pbet}
\alias{item_characteristics_sampler_pbet}
\title{Sampler for PBET}
\usage{
item_characteristics_sampler_pbet(
  pars = list(key_easy = 5L, key_hard = 5L, melody_length = 5:18)
)
}
\arguments{
\item{pars}{}
}
\description{
Sampler for PBET
}
