% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBET.R
\name{PBET}
\alias{PBET}
\title{Deploy the PBET}
\usage{
PBET(
  aws_credentials,
  num_items = list(find_the_note = 6L, arrhythmic = 10L, rhythmic = 10L),
  item_bank = itembankr::WJD,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url,
  examples = 2,
  final_results = TRUE
)
}
\arguments{
\item{final_results}{}
}
\value{

}
\description{
Deploy the PBET
}
