% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WJD.R
\docType{data}
\name{WJD_ngram_minified}
\alias{WJD_ngram_minified}
\title{Weimar Jazz Database (WJD), N-gram version (minified)}
\format{
A function containing different versions of the WJD data frame, returned with "main", "ngram", "phrases" and "files" keys
}
\source{
\url{https://jazzomat.hfm-weimar.de/dbformat/dboverview.html}
}
\usage{
WJD_ngram_minified
}
\description{
A dataset containing the melodies and other attributes of just over 11,000 melodic jazz solo phrases
}
\keyword{datasets}
