% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBET.R
\name{PBET}
\alias{PBET}
\title{Deploy the PBET}
\usage{
PBET(
  num_items = list(interval_perception = 0L, find_this_note = 0L, arrhythmic =
    list(key_easy = 5L, key_hard = 5L), rhythmic = list(key_easy = 5L, key_hard = 5L),
    wjd_audio = list(easy = 0L, hard = 0L)),
  melody_length = 3:15,
  item_bank = WJD::WJD,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url = "https://adaptiveeartraining.com",
  examples = list(find_this_note = 2L, arrhythmic = list(easy = 1L, hard = 1L),
    rhythmic = list(easy = 0L, hard = 0L), wjd_audio = list(easy = 0L, hard = 0L)),
  final_results = TRUE,
  musicassessr_state = "production",
  item_characteristics_sampler_function = item_characteristics_sampler_pbet,
  get_trial_characteristics_function = get_trial_characteristics_pbet,
  max_goes_forced = FALSE,
  max_goes = 3L,
  local_app_file_dir = "/Users/sebsilas/aws-musicassessr-local-file-upload/files/",
  sonic_annotator_local_location = "/Users/sebsilas/sonic-annotator",
  test_username = character(),
  store_results_in_db = FALSE,
  get_answer_midi = musicassessr::get_answer_midi,
  get_answer_audio = musicassessr::get_answer_pyin_melodic_production,
  give_first_melody_note = TRUE,
  with_final_page = TRUE,
  show_socials = TRUE,
  gold_msi = TRUE,
  headphones_test = TRUE,
  get_user_info = TRUE,
  microphone_test = TRUE,
  copy_audio_to_location = NULL
)
}
\arguments{
\item{copy_audio_to_location}{}
}
\value{

}
\description{
Deploy the PBET
}
