% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBET.R
\name{PBET_standalone}
\alias{PBET_standalone}
\title{Deploy the PBET as a standalone test}
\usage{
PBET_standalone(
  num_items = list(interval_perception = 24L, find_this_note = 6L, arrhythmic =
    list(key_easy = 5L, key_hard = 5L), rhythmic = list(key_easy = 5L, key_hard = 5L),
    wjd_audio = list(key_easy = 5L, key_hard = 5L)),
  melody_length = c(3, 15),
  item_bank = WJD::WJD,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url = character(),
  examples = list(find_this_note = 2L, interval_perception = 2L, arrhythmic = list(easy
    = 1L, hard = 1L), rhythmic = list(easy = 1L, hard = 1L), wjd_audio = 0L),
  final_results = TRUE,
  musicassessr_aws = FALSE,
  item_characteristics_sampler_function = item_characteristics_sampler_pbet,
  get_trial_characteristics_function = get_trial_characteristics_pbet,
  max_goes_forced = FALSE,
  max_goes = 4L,
  test_username = character(),
  store_results_in_db = FALSE,
  get_answer_function_midi = musicassessr::get_answer_midi_melodic_production,
  get_answer_function_audio = musicassessr::get_answer_pyin_melodic_production,
  give_first_melody_note = TRUE,
  with_final_page = TRUE,
  show_socials = TRUE,
  gold_msi = TRUE,
  headphones_test = TRUE,
  get_user_info = TRUE,
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  append_trial_block_before = psychTestR::module("before"),
  append_trial_block_after = psychTestR::module("after"),
  stop_recording_after = 30,
  skip_setup = FALSE,
  concise_wording = FALSE,
  app_name = "",
  get_self_chosen_anonymous_id = FALSE,
  adjust_range = TRUE,
  main_module_name = "PBET",
  input = c("midi_keyboard_or_microphone", "microphone", "midi_keyboard"),
  sampler_function_arrhythmic = musicassessr::sample_arrhythmic,
  sampler_function_rhythmic = musicassessr::sample_rhythmic,
  ...
)
}
\arguments{
\item{sampler_function_rhythmic}{}
}
\value{

}
\description{
Deploy the PBET as a standalone test
}
