% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBET.R
\name{PBET_standalone}
\alias{PBET_standalone}
\title{Deploy the PBET as a standalone test}
\usage{
PBET_standalone(
  num_items = list(interval_perception = 24L, find_the_note = 6L, arrhythmic =
    list(key_easy = 5L, key_hard = 5L), rhythmic = list(key_easy = 5L, key_hard = 5L),
    wjd_audio = list(key_easy = 5L, key_hard = 5L)),
  melody_length = 3:15,
  item_bank = itembankr::WJD,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url = "https://adaptiveeartraining.com",
  examples = list(easy = 1L, hard = 1L),
  final_results = TRUE,
  musicassessr_state = "production",
  item_characteristics_sampler_function = item_characteristics_sampler_pbet,
  get_trial_characteristics_function = get_trial_characteristics_pbet
)
}
\arguments{
\item{get_trial_characteristics_function}{}
}
\value{

}
\description{
Deploy the PBET as a standalone test
}
