% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBET.R
\name{PBET_standalone}
\alias{PBET_standalone}
\title{Deploy the PBET as a standalone test}
\usage{
PBET_standalone(
  num_items = list(interval_perception = 24L, find_this_note = 6L, arrhythmic =
    list(key_easy = 5L, key_hard = 5L), rhythmic = list(key_easy = 5L, key_hard = 5L),
    wjd_audio = list(key_easy = 5L, key_hard = 5L)),
  melody_length = 5:18,
  arrhythmic_item_bank = PBET::WJD_phrase_item_bank,
  rhythmic_item_bank = PBET::WJD_phrase_item_bank,
  demographics = TRUE,
  demo = FALSE,
  feedback = FALSE,
  admin_password = "demo",
  SNR_test = TRUE,
  get_range = TRUE,
  absolute_url = character(),
  num_examples = list(find_this_note = 2L, interval_perception = 2L, arrhythmic =
    list(key_easy = 1L, key_hard = 1L), rhythmic = list(key_easy = 1L, key_hard = 1L),
    wjd_audio = 0L),
  final_results = FALSE,
  item_characteristics_sampler_function = item_characteristics_sampler_pbet,
  get_trial_characteristics_function = get_trial_characteristics_pbet,
  max_goes_forced = FALSE,
  max_goes = 4L,
  test_username = character(),
  get_answer_function_midi = musicassessr::get_answer_midi_melodic_production,
  get_answer_function_audio = musicassessr::get_answer_pyin_melodic_production,
  give_first_melody_note = TRUE,
  with_final_page = TRUE,
  show_socials = TRUE,
  gold_msi = TRUE,
  headphones_test = TRUE,
  get_user_info = TRUE,
  microphone_test = TRUE,
  allow_repeat_SNR_tests = TRUE,
  append_trial_block_before = psychTestR::module("before"),
  append_trial_block_after = psychTestR::module("after"),
  stop_recording_after = 30,
  skip_setup = FALSE,
  concise_wording = FALSE,
  app_name = "",
  get_self_chosen_anonymous_id = FALSE,
  adjust_range = TRUE,
  main_module_name = "PBET",
  input_type = c("midi_keyboard_or_microphone", "microphone", "midi_keyboard"),
  sampler_function_arrhythmic = musicassessr::sample_arrhythmic,
  sampler_function_rhythmic = musicassessr::sample_rhythmic,
  test_name = "Play By Ear Test",
  learn_test_paradigm = FALSE,
  rel_to_abs_mel_function = musicassessr::transpose_melody_to_easy_or_hard_key,
  add_consent_form = FALSE,
  proportion_visual = list(learn = 0, test = 0),
  default_range = musicassessr::set_default_range("Piano"),
  default_singing_range = NULL,
  melody_block_paradigm = c("standard", "sing_melody_first",
    "learn_phase_visual_display_modality"),
  show_introduction = TRUE,
  num_items_review = list(arrhythmic = 0L, rhythmic = 0L),
  instrument_id = NULL,
  user_id = NULL,
  experiment_id = NULL,
  present_continue_to_new_test_page = TRUE,
  asynchronous_api_mode = FALSE,
  left_margin = 1L,
  right_margin = 1L,
  css = system.file("www/css/musicassessr.css", package = "musicassessr"),
  content_border = "solid 3px #bfd5d9",
  full_screen = FALSE,
  show_instructions = TRUE,
  presampled_item_bank = FALSE,
  mute_midi_playback = FALSE,
  ...
)
}
\arguments{
\item{instrument_id}{Set what instrument is being used a priori.}

\item{mute_midi_playback}{}
}
\description{
Deploy the PBET as a standalone test
}
